Create APOORVA_13_AUG_2023.ipynb
You have been given the role of a business analyst for an E-Commerce company and have been asked to prepare a basic report on the data. Here, we are analysing different aspects of E-Commerce Company by doing Exploratory Data Analysis. The questions and the respective points per question are mentioned in the Jupyter Notebook shared separately.